[APPREQ-629][QC-318][QC-320] fixed SAVE x% labels on search results text view so that it will not be shown if tiered sale is less than 10% refined tiered price sale out of bounds logging
git-svn-id: https://appdevsvn.nj01/appdev/FreshDirect/trunk@5171 842219ff-9341-0410-a943-573460bdd4de

+++ projects/ERPServices/src/com/freshdirect/common/pricing/util/DealsHelper.java
@@ -116,7 +116,7 @@ public class DealsHelper {
 		return highest > 0 ? highest : -1;
 	}
 	
-	public static boolean isDealOutOfBounds(double price) {
+	public static boolean isDealOutOfBounds(int price) {
 		return price < FDStoreProperties.getDealsLowerLimit()
 				|| price > FDStoreProperties.getDealsUpperLimit();
 	}
